refactor: Change comparison file names.
