+++ src/pages/Comparison/components/EachComparisonOptions.tsx
