+++ src/pages/Comparison/components/PageTitle.tsx
@@ -14,7 +14,7 @@ const PageTitle = ({ title, backUrl, backUrlTitle }: PageTitleType) => {
       <div>
         <h1 className='text-2xl font-bold'>{title}</h1>
       </div>
-      <Link className='flex items-center gap-2 text-muted-foreground' to={backUrl}>
+      <Link className='flex items-center gap-2 text-muted-foreground w-52' to={backUrl}>
         <ArrowLeftIcon className='w-4 h-4' />
         {backUrlTitle}
       </Link>
