+++ src/pages/Comparison/components/Scenario/EachScenarioOptions.tsx
